modifies check_boston-numbers
